refactor(actor): extract aether-actor SDK from aether-component (issue 509)
Phase 1 of ADR-0074. Lift the actor SDK out of `aether-component` into
a new `aether-actor` crate so the same types and helpers can carry
both the wasm-guest path and the future native-capability path.

The new `MailTransport` trait (transport.rs) is the split point —
five associated functions matching the `_p32` FFI surface. `Mail`,
`Ctx`, `Sink`, `InitCtx`, `DropCtx`, `Slot`, `KindId`, `WaitError`,
the `wait_reply` helper, and the typed-handle module move into
`aether-actor`. The transport-coupled types (`Sink<K, T>`,
`Ctx<'_, T>`, etc.) are generic over `T: MailTransport`; method
bodies dispatch through `T::*` instead of `unsafe { raw::* }`.

`aether-component` keeps the wasm-specific shim: `raw.rs`, the new
`WasmTransport` ZST that delegates to `raw::*`, the `Component`
trait (specialised to `WasmTransport` via 1-arg type aliases for
`Sink<K>` / `Ctx<'a>` / `InitCtx<'a>` / `DropCtx<'a>` / `Handle<K>`),
the `export!` macro, and the wasm-flavoured `io` / `net` / `log`
helper modules. A 1-arg `resolve_sink<K>` facade preserves the
existing call shape so component code compiles unchanged.

`#[handlers]` macro paths through `::aether_component::*` resolve
through the type aliases — no macro change needed in this phase.
The macro genericity work moves to Phase 2 alongside `NativeTransport`.

Closes nothing on its own — issue 509 stays open until the remaining
four phases ship (NativeTransport + capability migration, render
collapse, host-side wait_reply + cross-class guard, wire rename).
